// Front/Frontend/src/app/swagger/services/models/demande-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface DemandeDto {
  id:number;
  categorie?: string;
  impact?: string;
  lieu?: string;
  urgence?: string;
  employeeMatricule? : string;
  description? : string;
  pieceJointe? : string;
  type? : string;
  status?: string;
  createdAt?: string;
  updatedAt?: string;
  createur?:String;

  

}